+++ server/graphql/generated.ts
@@ -1575,6 +1575,26 @@ export type GQLIntegrationNoInputCredentialsError = GQLError & {
   readonly type: ReadonlyArray<Scalars['String']['output']>;
 };
 
+export type GQLInvalidItemTypeHiddenFieldsError = GQLError & {
+  readonly __typename?: 'InvalidItemTypeHiddenFieldsError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
+};
+
+export type GQLInvalidItemTypeSchemaError = GQLError & {
+  readonly __typename?: 'InvalidItemTypeSchemaError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
+};
+
 export type GQLInvalidateReportsFromReporterInput = {
   /**
    * Scopes the sweep to a single MRT job. When omitted, every pending job
@@ -1746,6 +1766,16 @@ export type GQLItemTypeIdentifierInput = {
 
 export type GQLItemTypeNameAlreadyExistsError = GQLError & {
   readonly __typename?: 'ItemTypeNameAlreadyExistsError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
+};
+
+export type GQLItemTypeSchemaIncompatibleError = GQLError & {
+  readonly __typename?: 'ItemTypeSchemaIncompatibleError';
   readonly detail?: Maybe<Scalars['String']['output']>;
   readonly pointer?: Maybe<Scalars['String']['output']>;
   readonly requestId?: Maybe<Scalars['String']['output']>;
@@ -2484,7 +2514,12 @@ export type GQLMutateBankResponse = {
 };
 
 export type GQLMutateContentItemTypeResponse =
-  GQLItemTypeNameAlreadyExistsError | GQLMutateContentTypeSuccessResponse;
+  | GQLInvalidItemTypeHiddenFieldsError
+  | GQLInvalidItemTypeSchemaError
+  | GQLItemTypeNameAlreadyExistsError
+  | GQLItemTypeSchemaIncompatibleError
+  | GQLMutateContentTypeSuccessResponse
+  | GQLNotFoundError;
 
 export type GQLMutateContentRuleSuccessResponse = {
   readonly __typename?: 'MutateContentRuleSuccessResponse';
@@ -2534,7 +2569,12 @@ export type GQLMutateRoutingRulesOrderSuccessResponse = {
 };
 
 export type GQLMutateThreadItemTypeResponse =
-  GQLItemTypeNameAlreadyExistsError | GQLMutateThreadTypeSuccessResponse;
+  | GQLInvalidItemTypeHiddenFieldsError
+  | GQLInvalidItemTypeSchemaError
+  | GQLItemTypeNameAlreadyExistsError
+  | GQLItemTypeSchemaIncompatibleError
+  | GQLMutateThreadTypeSuccessResponse
+  | GQLNotFoundError;
 
 export type GQLMutateThreadTypeSuccessResponse = {
   readonly __typename?: 'MutateThreadTypeSuccessResponse';
@@ -2542,7 +2582,12 @@ export type GQLMutateThreadTypeSuccessResponse = {
 };
 
 export type GQLMutateUserItemTypeResponse =
-  GQLItemTypeNameAlreadyExistsError | GQLMutateUserTypeSuccessResponse;
+  | GQLInvalidItemTypeHiddenFieldsError
+  | GQLInvalidItemTypeSchemaError
+  | GQLItemTypeNameAlreadyExistsError
+  | GQLItemTypeSchemaIncompatibleError
+  | GQLMutateUserTypeSuccessResponse
+  | GQLNotFoundError;
 
 export type GQLMutateUserRuleSuccessResponse = {
   readonly __typename?: 'MutateUserRuleSuccessResponse';
@@ -5619,10 +5664,14 @@ export type GQLResolversUnionTypes<_RefType extends Record<string, unknown>> = {
         data: _RefType['CustomAction'];
       });
   MutateContentItemTypeResponse:
+    | GQLInvalidItemTypeHiddenFieldsError
+    | GQLInvalidItemTypeSchemaError
     | GQLItemTypeNameAlreadyExistsError
+    | GQLItemTypeSchemaIncompatibleError
     | (Omit<GQLMutateContentTypeSuccessResponse, 'data'> & {
         data?: Maybe<_RefType['ContentItemType']>;
-      });
+      })
+    | GQLNotFoundError;
   MutateHashBankResponse:
     | GQLMatchingBankNameExistsError
     | (Omit<GQLMutateHashBankSuccessResponse, 'data'> & {
@@ -5634,15 +5683,23 @@ export type GQLResolversUnionTypes<_RefType extends Record<string, unknown>> = {
         data: _RefType['LocationBank'];
       });
   MutateThreadItemTypeResponse:
+    | GQLInvalidItemTypeHiddenFieldsError
+    | GQLInvalidItemTypeSchemaError
     | GQLItemTypeNameAlreadyExistsError
+    | GQLItemTypeSchemaIncompatibleError
     | (Omit<GQLMutateThreadTypeSuccessResponse, 'data'> & {
         data?: Maybe<_RefType['ThreadItemType']>;
-      });
+      })
+    | GQLNotFoundError;
   MutateUserItemTypeResponse:
+    | GQLInvalidItemTypeHiddenFieldsError
+    | GQLInvalidItemTypeSchemaError
     | GQLItemTypeNameAlreadyExistsError
+    | GQLItemTypeSchemaIncompatibleError
     | (Omit<GQLMutateUserTypeSuccessResponse, 'data'> & {
         data?: Maybe<_RefType['UserItemType']>;
-      });
+      })
+    | GQLNotFoundError;
   PartialItemsResponse:
     | GQLPartialItemsEndpointResponseError
     | GQLPartialItemsInvalidResponseError
@@ -5746,9 +5803,12 @@ export type GQLResolversInterfaceTypes<
     | GQLIntegrationConfigUnsupportedIntegrationError
     | GQLIntegrationEmptyInputCredentialsError
     | GQLIntegrationNoInputCredentialsError
+    | GQLInvalidItemTypeHiddenFieldsError
+    | GQLInvalidItemTypeSchemaError
     | GQLInviteUserTokenExpiredError
     | GQLInviteUserTokenMissingError
     | GQLItemTypeNameAlreadyExistsError
+    | GQLItemTypeSchemaIncompatibleError
     | GQLJobHasAlreadyBeenSubmittedError
     | GQLLocationBankNameExistsError
     | GQLLoginIncorrectPasswordError
@@ -6077,6 +6137,8 @@ export type GQLResolversTypes = {
   IntegrationEmptyInputCredentialsError: ResolverTypeWrapper<GQLIntegrationEmptyInputCredentialsError>;
   IntegrationMetadata: ResolverTypeWrapper<GQLIntegrationMetadata>;
   IntegrationNoInputCredentialsError: ResolverTypeWrapper<GQLIntegrationNoInputCredentialsError>;
+  InvalidItemTypeHiddenFieldsError: ResolverTypeWrapper<GQLInvalidItemTypeHiddenFieldsError>;
+  InvalidItemTypeSchemaError: ResolverTypeWrapper<GQLInvalidItemTypeSchemaError>;
   InvalidateReportsFromReporterInput: GQLInvalidateReportsFromReporterInput;
   InvalidateReportsFromReporterSuccessResponse: ResolverTypeWrapper<GQLInvalidateReportsFromReporterSuccessResponse>;
   InviteUserInput: GQLInviteUserInput;
@@ -6120,6 +6182,7 @@ export type GQLResolversTypes = {
   >;
   ItemTypeIdentifierInput: GQLItemTypeIdentifierInput;
   ItemTypeNameAlreadyExistsError: ResolverTypeWrapper<GQLItemTypeNameAlreadyExistsError>;
+  ItemTypeSchemaIncompatibleError: ResolverTypeWrapper<GQLItemTypeSchemaIncompatibleError>;
   ItemTypeSchemaVariant: ResolverTypeWrapper<ItemTypeSchemaVariantResolverValue>;
   ItemTypeSchemaVariantInput: ResolverTypeWrapper<ItemTypeSchemaVariantInputResolverValue>;
   ItemWithParents: ResolverTypeWrapper<
@@ -6869,6 +6932,8 @@ export type GQLResolversParentTypes = {
   IntegrationEmptyInputCredentialsError: GQLIntegrationEmptyInputCredentialsError;
   IntegrationMetadata: GQLIntegrationMetadata;
   IntegrationNoInputCredentialsError: GQLIntegrationNoInputCredentialsError;
+  InvalidItemTypeHiddenFieldsError: GQLInvalidItemTypeHiddenFieldsError;
+  InvalidItemTypeSchemaError: GQLInvalidItemTypeSchemaError;
   InvalidateReportsFromReporterInput: GQLInvalidateReportsFromReporterInput;
   InvalidateReportsFromReporterSuccessResponse: GQLInvalidateReportsFromReporterSuccessResponse;
   InviteUserInput: GQLInviteUserInput;
@@ -6901,6 +6966,7 @@ export type GQLResolversParentTypes = {
   };
   ItemTypeIdentifierInput: GQLItemTypeIdentifierInput;
   ItemTypeNameAlreadyExistsError: GQLItemTypeNameAlreadyExistsError;
+  ItemTypeSchemaIncompatibleError: GQLItemTypeSchemaIncompatibleError;
   ItemWithParents: Omit<GQLItemWithParents, 'item' | 'parents'> & {
     item: GQLResolversParentTypes['ItemSubmissions'];
     parents: ReadonlyArray<GQLResolversParentTypes['ItemSubmissions']>;
@@ -9013,9 +9079,12 @@ export type GQLErrorResolvers<
     | 'IntegrationConfigUnsupportedIntegrationError'
     | 'IntegrationEmptyInputCredentialsError'
     | 'IntegrationNoInputCredentialsError'
+    | 'InvalidItemTypeHiddenFieldsError'
+    | 'InvalidItemTypeSchemaError'
     | 'InviteUserTokenExpiredError'
     | 'InviteUserTokenMissingError'
     | 'ItemTypeNameAlreadyExistsError'
+    | 'ItemTypeSchemaIncompatibleError'
     | 'JobHasAlreadyBeenSubmittedError'
     | 'LocationBankNameExistsError'
     | 'LoginIncorrectPasswordError'
@@ -9560,6 +9629,67 @@ export type GQLIntegrationNoInputCredentialsErrorResolvers<
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 };
 
+export type GQLInvalidItemTypeHiddenFieldsErrorResolvers<
+  ContextType = Context,
+  ParentType extends
+    GQLResolversParentTypes['InvalidItemTypeHiddenFieldsError'] =
+    GQLResolversParentTypes['InvalidItemTypeHiddenFieldsError'],
+> = {
+  detail?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  pointer?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  requestId?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  status?: Resolver<GQLResolversTypes['Int'], ParentType, ContextType>;
+  title?: Resolver<GQLResolversTypes['String'], ParentType, ContextType>;
+  type?: Resolver<
+    ReadonlyArray<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type GQLInvalidItemTypeSchemaErrorResolvers<
+  ContextType = Context,
+  ParentType extends GQLResolversParentTypes['InvalidItemTypeSchemaError'] =
+    GQLResolversParentTypes['InvalidItemTypeSchemaError'],
+> = {
+  detail?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  pointer?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  requestId?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  status?: Resolver<GQLResolversTypes['Int'], ParentType, ContextType>;
+  title?: Resolver<GQLResolversTypes['String'], ParentType, ContextType>;
+  type?: Resolver<
+    ReadonlyArray<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
 export type GQLInvalidateReportsFromReporterSuccessResponseResolvers<
   ContextType = Context,
   ParentType extends
@@ -9843,6 +9973,37 @@ export type GQLItemTypeNameAlreadyExistsErrorResolvers<
   ContextType = Context,
   ParentType extends GQLResolversParentTypes['ItemTypeNameAlreadyExistsError'] =
     GQLResolversParentTypes['ItemTypeNameAlreadyExistsError'],
+> = {
+  detail?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  pointer?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  requestId?: Resolver<
+    Maybe<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  status?: Resolver<GQLResolversTypes['Int'], ParentType, ContextType>;
+  title?: Resolver<GQLResolversTypes['String'], ParentType, ContextType>;
+  type?: Resolver<
+    ReadonlyArray<GQLResolversTypes['String']>,
+    ParentType,
+    ContextType
+  >;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type GQLItemTypeSchemaIncompatibleErrorResolvers<
+  ContextType = Context,
+  ParentType extends
+    GQLResolversParentTypes['ItemTypeSchemaIncompatibleError'] =
+    GQLResolversParentTypes['ItemTypeSchemaIncompatibleError'],
 > = {
   detail?: Resolver<
     Maybe<GQLResolversTypes['String']>,
@@ -10871,7 +11032,12 @@ export type GQLMutateContentItemTypeResponseResolvers<
     GQLResolversParentTypes['MutateContentItemTypeResponse'],
 > = {
   __resolveType: TypeResolveFn<
-    'ItemTypeNameAlreadyExistsError' | 'MutateContentTypeSuccessResponse',
+    | 'InvalidItemTypeHiddenFieldsError'
+    | 'InvalidItemTypeSchemaError'
+    | 'ItemTypeNameAlreadyExistsError'
+    | 'ItemTypeSchemaIncompatibleError'
+    | 'MutateContentTypeSuccessResponse'
+    | 'NotFoundError',
     ParentType,
     ContextType
   >;
@@ -11003,7 +11169,12 @@ export type GQLMutateThreadItemTypeResponseResolvers<
     GQLResolversParentTypes['MutateThreadItemTypeResponse'],
 > = {
   __resolveType: TypeResolveFn<
-    'ItemTypeNameAlreadyExistsError' | 'MutateThreadTypeSuccessResponse',
+    | 'InvalidItemTypeHiddenFieldsError'
+    | 'InvalidItemTypeSchemaError'
+    | 'ItemTypeNameAlreadyExistsError'
+    | 'ItemTypeSchemaIncompatibleError'
+    | 'MutateThreadTypeSuccessResponse'
+    | 'NotFoundError',
     ParentType,
     ContextType
   >;
@@ -11029,7 +11200,12 @@ export type GQLMutateUserItemTypeResponseResolvers<
     GQLResolversParentTypes['MutateUserItemTypeResponse'],
 > = {
   __resolveType: TypeResolveFn<
-    'ItemTypeNameAlreadyExistsError' | 'MutateUserTypeSuccessResponse',
+    | 'InvalidItemTypeHiddenFieldsError'
+    | 'InvalidItemTypeSchemaError'
+    | 'ItemTypeNameAlreadyExistsError'
+    | 'ItemTypeSchemaIncompatibleError'
+    | 'MutateUserTypeSuccessResponse'
+    | 'NotFoundError',
     ParentType,
     ContextType
   >;
@@ -15299,6 +15475,8 @@ export type GQLResolvers<ContextType = Context> = {
   IntegrationEmptyInputCredentialsError?: GQLIntegrationEmptyInputCredentialsErrorResolvers<ContextType>;
   IntegrationMetadata?: GQLIntegrationMetadataResolvers<ContextType>;
   IntegrationNoInputCredentialsError?: GQLIntegrationNoInputCredentialsErrorResolvers<ContextType>;
+  InvalidItemTypeHiddenFieldsError?: GQLInvalidItemTypeHiddenFieldsErrorResolvers<ContextType>;
+  InvalidItemTypeSchemaError?: GQLInvalidItemTypeSchemaErrorResolvers<ContextType>;
   InvalidateReportsFromReporterSuccessResponse?: GQLInvalidateReportsFromReporterSuccessResponseResolvers<ContextType>;
   InviteUserToken?: GQLInviteUserTokenResolvers<ContextType>;
   InviteUserTokenExpiredError?: GQLInviteUserTokenExpiredErrorResolvers<ContextType>;
@@ -15317,6 +15495,7 @@ export type GQLResolvers<ContextType = Context> = {
   ItemTypeBase?: GQLItemTypeBaseResolvers<ContextType>;
   ItemTypeIdentifier?: GQLItemTypeIdentifierResolvers<ContextType>;
   ItemTypeNameAlreadyExistsError?: GQLItemTypeNameAlreadyExistsErrorResolvers<ContextType>;
+  ItemTypeSchemaIncompatibleError?: GQLItemTypeSchemaIncompatibleErrorResolvers<ContextType>;
   ItemTypeSchemaVariant?: GQLItemTypeSchemaVariantResolvers;
   ItemTypeSchemaVariantInput?: GQLItemTypeSchemaVariantInputResolvers;
   ItemWithParents?: GQLItemWithParentsResolvers<ContextType>;